Added integration test for app.CheckConfig() and moved it in package configdb This function was only tested with mocks, and the test was very difficult to adapt with new changes It now has an integration test with database, which guarentees it works and will make it easier to change in the future The function CheckConfig() has to be moved in another package: configdb, because of a circular dependency issue: app -> testhelpers -> app The unit test with mocks TestCheckConfig can be removed, we'll plan that just after we're sure that it passes the CI
